Move Chart Bar to new architecture
