+++ src/components/Dashboard/SimpleHorizontalBarChart.tsx
